Luks open (#8)
* Add yubikey-luks-open script

It allows for opening LUKS containers protected with yubikey-luks outside initramfs. It's useful for external encrypted disks or system rescue in case of broken initramfs.

* yubikey-luks-open script: add to debian build

* Readme: add yubikey-luks-open info

* intendation fix

* yubikey-luks-open: Adjust wording for avalaible option

* yubikey-luks-open: get rid of tmp keyfile

* yubikey-luks-enroll: get rid of tmp keyfile

* yubikey-luks-enroll: fix indedation

* Revert: yubikey-luks-enroll: get rid of tmp keyfile

In this case we actually need tmp file. Sorry for the noise.

*  yubikey-luks-enroll: get rid of tmp keyfile V2

I finally found a clean way to get rid of tmp file in yubikey-luks-enroll too. No more tmpfiles :)